Made another big modification.
So this one switches things up a bit.

Firstly, activation conditions now affect cost a bit.

Secondly, some costs have been tweaked again, and a new effect has
been added.

Thirdly, and most importantly, target selection no longer in and of
itself affects cost, but rather it is incorporated into the cost of
effects that utilize targets. This allows for a mixture of both
methods I experimented with (refining the way costs are defined),
and it also cleans up the card texts to avoid that redundancy of,
"Target up to 2 cards. Destroy up to 2 targets." Now, it would say,
"Target up to 2 cards. Destroy all targets." Of course, the one
downside is that all cards had to be updated.

I'm hopeful that this will be the last change of this nature.