Introduced revision, revised the card exhaustion cost to creature exhaustion.
I found in my playtests that it's common for it to be prudent to exhaust
traps for cards using this cost, and this made the cost far too
useful. By forcing players to exhaust specifically creatures when
using these cards, that should balance it out.

Note: this begins the 2nd edition development.