Redacted the direct damage effect.
This effect, the more that I think about it, is very likely to be
an effect that makes the game less fun, both because they are
difficult to counter (you can't defend with creatures) and because
they eliminate the suspense of dice-rolling. As such, I have decided
to redact it.