Added another dice re-rolling effect for use with your own cards.
The first dice re-rolling effect only works in a chain, but you can't
chain to your own cards. So this new effect only workd *after* the
current chain for the remainder of the turn, enabling you to use it
with your own cards by playing a card that does it in advance.