Added tip cards and a retaliation rule.
Tip cards are there to show the basic rules of the game without
having to have the whole manual out.

The new retaliation rule is: if you're attacking the player, then
any of the defending team's remaining dice get applied to you as
damage. The main reason for doing this is to minimize the initial
advantage the starting player has, but it should also have the
effect of making targeting decisions more interesting.