Added "destroy" costs.
These costs are designed to be used for combos in decks that focus
on such things as milling one's own deck or discarding one's own cards.
This way, you can destroy your own cards without having to pay extra
for it, and it will also give you some cost relief (but not as much
as sacrificing).