Changed the deck size from strictly 50 to between 40 and 60.
I realized that 50 is not divisible by 8 (the number of cards per set),
so I checked, and in fact, 40-card decks are perfectly capable of
lasting until the end without decking out. Also, being able to vary
the deck size means you can better prepare for self-milling strategies
and the like, so a variable amount of possible cards in a deck is probably
a good idea.