Fixes plus lots of new example cards.
I'm putting together a snake archetype of sorts. It's for myself,
but they might as well be included as example cards too, eh? ;)

The snakes are based on a strategy of flooding the opponent's
creatures with damage counters. Since damage counter placement is
so expensive, this also includes flooding the field with snakes,
probably either by summoning them from the deck or graveyyard or by
drawing lots of cards.