Added a rule prohibiting attacking the same target multiple times in one turn.
This rule is meant to prevent large armies of weak creatures from
getting past small armies of strong creatures just by attacking with
one at a time. Now, for example, you can't attack the player 10 times
with a really weak enemy and count on the fact that they can only
defend twice; instead, you have to either attack with all at once,
or perhaps target one of the potential defenders with some of the
army and then the player with the rest of the army, for example.