Added a first-turn caveat: only place one or draw one.
I decided to add this when it occurred to me that 7jeronimo7's
OTK strategy hinges on a first-turn advantage, so my previous
assumption that this advantage is insignificant is incorrect.
So this is a simple nerf to nullify it: the very first player, on
the very first turn, only gets to place one resource OR draw one
card.