Changed layout of sets to accompany larger printer margins.
Half-inch margins are quite small, and the benefit of paper saved
is pretty slim making the margin 0.5 inches instead of 0.75 inches.
My printer can handle the latter, but not the former, and I'd hazard
to guess that other people's printers are the same story.

Tip cards are being kept as-is, at least for now, because those don't
need to be sized exactly.