Switched up the set arrangement again.
After some frustration and experimentation, I found that *any*
print of cards from the PDF caused the cards to be smaller! So
screw it, I've changed it to a 3x3 arrangement, with half-inch
margins on the side and quarter-inch margins on the top and bottom.
They print perfectly on my printer from the ODG.

For that reason also, I've gotten rid of the PDFs.