Added limits, revised a card text.
So, first is the revision: the card text targeting a resource can
now only target your own resources. This is because I realized
(through playing with @7jeronimo7) that it's very easy to target
enemy resources with a creature ability over and over again to
make it completely impossible for them to play, since there's a
strict limit on how many resources can be played. So I changed the
resource targeting text to only allow targeting your own resources,
which shouldn't be a problem.

Second, limits. I've added support for "variable limits", i.e.
limits on what you're allowed to set variables to. Currently, I've
set a limit to the self-milling and hitpoint costs.