Added a rule preventing new chains from starting within effects.
I had to think about this quite a lot, but I ultimately decided that
you should not be allowed to chain to actions taken within effects
(such summoning of a creature due to an effect) because allowing this
is a major potential source of confusion. Suppose, for example, that
5 cards in a chain summon a creature. That would be five new chains that all
start simultaneously, which would be very hard to keep track of and
it's unclear what the order of the chains should be. Overall, it just
doesn't seem to be worth it.