Archetypes experiment
I'm not sure about this, but I'm resisting the urge to purge the
whole thing, just in case it's a good idea after all. This is an
idea I've been throwing around in my head for some time now: special
"archetype" cards, which would allow you to create cards that work
well together. The problem I'm facing is how to make it worthwhile
from a gameplay perspective without also making it game-breaking.
It could be challenging, to say the least.