Finally came up with a viable option to implement "archetypes" of a sort.
Introducing Super Cards! Super Cards are just like regular cards,
but are allowed to have certain ability texts that other cards are
not. In general, these are texts which are either extremely
powerful, or more commonly, work together with other Super Cards.
The mechanism used to prevent making this mechanic game-breaking is
to limit the number of Super Cards allowed in a deck to 25%, so e.g.
no more than 10 in a 40-card deck and no more than 15 in a 60-card
deck.

Needs testing and possibly tweaking. Who knows? This might not work
out at all. 7th Edition is quite solid, so I don't want to rush
such a radical change before testing it.