Added the "reserves" rule change.
This is an idea that we tested over the last couple of months, and
I am overall pleased with the results. The reserves make it much
easier to put together your engine (by allowing key cards to be
put in the reserves instead of having to wait to draw them), and it
also makes the game much shorter by limiting deck size. Hitpoints
have also been reduced to 20 so that a proper win via hitpoints
is still the most likely outcome.