Added support for cost adjustments, plus some cost adjustments as well.
I've been thinking for months about how to support this, and what
I came up with is simply to note the edition a card was designed for
on the card.

I also made the first ever cost adjustments.

Just to make sure I have no trouble later on, I decided to go through
and insert the Edition Indicator into every example card. It was
long and tedious, but now it's done.