Must attack -> destroy if unexhausted at end of Battle Phase
This variant is much clearer and prevents potential edge cases
("I was going to attack your creature after I attacked you but the
creature is dead now!")

I could change this to "send to the graveyard" instead of "destroy",
but I think destruction is sufficient. It can be taken advantage of but
not in a game-breaking way. "Destroy 1 of your cards in play" would
be able to do the same thing with the same cost relief.