Converted the rulebook and card building guide to HTML.
This took a real long time to finish considering how much drudgery it
was, but I think was worth it. I'll be phasing out the ODF stuff in
favor of this new HTML stuff.

Something I did when transcribing was eliminate some redundant and
unnecessary information, notably the indication of what texts have
been revised in the card building guide.