Changed deck size to be 40-60 cards (instead of 30).
I don't remember if we did this back then in an effort to shorten
the game, but if we did, it was a bad way to keep the game shorter.
More likely the low deck size was an attempt to stop the old
"self-mill as cost" mechanic from being broken, but that didn't work
and we had to ban that a long time ago. I think it's fine (and even
beneficial) to increase the deck size and allow the player to choose
the exact deck size they want.