New keyword "Deal", new card "Celestial Dance" replacing "Consolation".
"Deal" replaces "All players draw". This effect has also been made
cheaper, since other than being an efficient deck-out strategy, it
basically gives an equal amount of advantage to everyone, which makes
it a pretty bad effect.

Also renamed the file for "Vow of Vengeance" to "Seeing Red" (which
we'll change the name to), thô we haven't actually changed the card
yet.