Let's make a massively breaking formula change.
The effect of activation conditions has always been rather
unsatisfying, and not done in a way that facilitates *beneficial*
activation conditions. So let's change that: activation conditions
now create a multiplier, just like target selections do. Of course,
that meant that the effect of activation conditions had to be
completely reselected, and any card from a previous edition that
used any activation condition has to be invalidated. Some tweaking
might be needed, but I think this'll be overall a much better design
than before.