Updated Holy Admission and added two new keywords.
"Down-level" is now used as a keyword for reducing level. Allows for
better clarification of rules around it in the rulebook and shortens
card text.

Also added "fatigue" as a keyword to replace the "doesn't get unexhausted
this turn" text. I also decided to strengthen that effect: instead of
only preventing the card being unexhausted, it also prevents all
stat changes, allowing some additional potential shenanigans.