Opened up the way cards are allowed to be constructed.
This allows chaining together multiple caveats and multiple "sub-abilities"
to make cards more complicated. The main purpose of this is to enable
a fusion mechanic which can also double as an easier way to design
your own cards, combining together abilities, caveats, and stats of
multiple cards. Unfortunately, this change requires the redaction of
the cost component which removes counters, as with this new rule change,
it would change from a downside of delaying an ability by a turn to a
practically obligatory math puzzle at card creation time to cut the cost
of large abilities as low as possible. The mechanic of removing counters
for cost isn't particularly fun to begin with, and therefore not worth
it.