Open up the gates to 0 cost creatures.
This doesn't affect any of the cards I've designed, since 0 cost
abilities still are impossible, but this creates the possibility of
weak, free creatures that can't attack or can't defend. It also makes
the cost calculations consistent with the fusion mechanic I want to
introduce.

This might warrant turning 12th Edition into another "clean slate"
release. Will have to think on this a bit.