Removed play restriction caveats and designated 12th as a clean slate.
Play restriction caveats I found to drag games out in testing, especially
with the fusion mechanic. They hinder rather than inviting creativity,
and so they have to go. Sorry, Berzerker. This is going to require a
huge number of cards to change.

I've also decided that given the changes to how costs are calculated,
a clean slate edition is justified.