Reinstate the minimum level of 1 for spells and abilities.
I had completely forgotten about the reason for this: that without
it, creatures would be incentivized to have virtually limitless
abilities, something which would be tedious and impractical. Level
0 creatures with no abilities are still possible, but abilities are
back to being level 1+ only. The cards which took advantage of the
level 0 ability rule have been adjusted accordingly.

Also removed a targeting text that I just noticed is incredibly
broken. (Balancing its cost would be possible, but would take away
the benefit of using it.)