Modified how abilities are constructed a bit.
This was mostly intended to streamline fusion creatures' ability
text, and it turned out to be way more complicated than I expected.
In short, there's a bit of a downside to this approach: by de-isolating
activation costs from the associated effects, it makes it much easier
to create a fusion creature that has a higher level than it should
have. However, it wasn't impossible before, doing this requires
using poorly-designed cards, and there's no gameplay advantage to it.
So I've decided I'm fine with it. I considered adding something more
complicated to really address it, but that level of complexity hurts
my brain and would probably take the fun out of playing.