Revamped activation conditions.
I've decided that it'd be best for the health of the game to get rid
of the whole cost modification effect of activation conditions. This
will simplify card cost calculations significantly and also help to
facilitate a "mutation" mechanic I'm thinking of implementing, since
the fusion mechanic didn't work.

Also added the keyword "Ghost-play" to clarify when cards can activate
their effects in special circumstances. I'm not 100% certain about
making the self-damage condition "Ghost-play"; have to check to make
sure there won't be an infinite loop. But if it doesn't create an
infinite loop of defense-gains, I'll want that, I think.