Replaced the deck-stack effect with a more powerful and expensive "bury".
Burying a card places it on the bottom of the deck. It's a purely
offensive move.

The reason for this change is deck-stacking was kind of caught between
a rock and a hard place. Bouncing was better for your own cards, but
destroying was better for opponents' cards, despite being cheaper
than deck-stacking. This should even the playing field a bit more.
The new bury effect is the closest thing to a permanent removal.