Completely removed caveats.
This was a hard decision to make, but I've come to the conclusion that
creature caveats don't really add to the creativity or depth of the
game, only to the complexity of it. All they really do is allow bigger
numbers on your creatures, and allow creatures to cosplay as Yu-Gi-Oh
cards.