Removed Enemy's Choice and similar texts.
I'm thinking that Enemy's Choice is detrimental as a mechanic to the
fun of the game, purely because of the economics of it. If it's better
cost-wise than the "standard" targetings and costs, it in practice
becomes the standard, and the meta favors obnoxious uses where you either
don't really care about the choice or force a choice which is favorable
to you.