Terminology change: creatures are now minions.
I mentioned this suggestion to someone as a joke, but honestly, I
love the sound of "minion" as a name for the creatures you command.
It sounds unique, like Yu-Gi-Oh's "monster".

I also got rid of an unimportant maximum offense/defense stat of 99.
I don't remember why that was put there, but it's honestly completely
unnecessary. There's no reason you'd want stats to be that high anyway,
and it just presents a needless complication in the rules.