Doubled the cost of down-leveling.
This is to prevent yet another numbers game, in this case one which
made infinite loops possible. The one I came up with was a 0 offense
and 0 defense creature with the following effect:

When this in-play card is destroyed (ghost-play): Target 1 in-play
card; destroy it. Target this card; down-level it by 9, then reveal
and deploy it, also place 1 Damage Counter on it.

With the previous pricing, this would leed to a level 9 creature which
can infinitely revive itself for free and destroy everything. With this
cost change, that creature would instead be level 14 and thus require
5 resources per loop instance.