Combined the exhaust and fatigue effects into one.
I like the fatigue effect as something distinct, with extra power
compared to an exhaustion counter, but it really should be intrinsically
combined with exhaustion rather than doing that combination manually.
This gets rid of the redundant, cheaper, and less flexible effect that
does the same thing as exhaustion counters.