Lowered the cost of down-leveling.
As long as the cost for a card to down-level itself is equal to the
level increase the effect brings, the cost is sufficient to prevent
down-leveling from being used for number games and can only be used
to "offset" a future cost with a present cost, e.g. by summoning a
minion or setting a trap that can down-level other cards.