Removed our special Resource Phase and replaced it with the Duel Masters way.
As time has gone on, I've started to dislike the KotC Resource Phase
and the effect it has on the game. It in particular makes card economy
very Yu-Gi-Oh-like, where having lots of cards in your hand lets you
sow resources faster. This was actually originally why the silo effect
was made the same cost as drawing; when you can choose to draw 2 or
sow 2, siloing really has a similar effect on the game compared to
drawing. By limiting it so that you can only sow 1 per turn, this
should mean that having tons of cards you don't need in your hand
can't be used to hyper-accelerate resource building, hopefully making
the game feel more fair.