Removed the definition of "up to" as a keyword and removed a bit from Excavate.
There was some text that said the "ability has no further effect",
which is actually inaccurate. That was something I figured it would
be a long time ago, but that was back before there was a visible
delineation between costs and effects, and copied from Yu-Gi-Oh.
Abilities now actually try to do as much as possible when e.g. there
are no valid targets.

The definition of "up to" which required targeting at least 1 card
was also a leftover, thô I don't particularly remember what it was
a leftover from. It's completely unnecessary; if there are zero
targets, then effects that do something to targets simply don't do
anything.