Added a mechanic known as "piercing damage".
This is meant to fight against stalling which can happen with 0 HP
minions. Now, 0 HP minions offer no protection to your HP, and in fact,
they create an inherent weak point that can be taken advantage of by
certain effects. This is an effort to keep games from stalling out so
much. Having tested this a bit, I find it to be a beneficial change,
albeit not quite enough.