Returned the edition number to 13. This will be 13th Edition, Revision 1.
I've decided that I don't want any additional changes, and with the
one change implemented being only a rulebook change that doesn't
affect card text at all, I don't think that justifies an increase
in edition number. No cards need to be reprinted, and the game is
basically the same.