Added "pawns", equivalent to monster tokens in Yu-Gi-Oh.
We decided to make it so that pawns are a specific level, but then
the player decides in the moment how to distribute SP and HP, which
should add an interesting dynamic.

Because it's not allowed to choose the number 0 in variables when
creating a card, it's not possible to create a pawn with a level
lower than 0. This is good, because otherwise it would be entirely
free to create an infinite army of level 0 pawns for sacrifice fodder.