Added emphasis to the requirement that variables must be non-zero.
There is now a card text that would be broken if someone were to
fail to notice the requirement that all variables must be set to
positive, non-zero integers. The previous phrasing was both in
another paragraph and not particularly emphasized. I've moved it
and added a strong tag to make it very difficult to miss.