Added a limited enemy resource targeting option.
I just realized one possible way to have this kind of interaction:
to restrict it to *exhausted* resources only. That means you have to
do something somewhat complicated to do something to enemy resources,
and even when you do, you guarantee (with the exception of siloed
resources) that the enemy gets to do at least one thing with said
resources, first.