Modified the costs of sheathing and siloing.
Sheathing cost has been increased to 3t, while siloing has been
decreased to 2t. This I think is more in-line with their usefulness;
siloing is roughly as good as drawing cards (faster but less flexible),
and sheathing is generally a more powerful removal than destruction
(dodges destruction procs and forces opponent to draw a card that
might be useless to them).