Added targeting protection in the form of Mirror Counters.
To balance these, I decided to make it so that it only forces the
player who used the ability to choose a different valid target, if
possible. This means that Mirror Counters will do nothing at all if
a minion is alone on the battlefield, and it also means that in most
cases, it will be easy for the player to choose a different minion
controlled by the same player, meaning the Mirror Counter can't
necessarily stop the damage entirely, only redirect it.