New idea: Ultra Points.
Copying from some other card games, Ultra Points are a points pool
limiting the number of times a certain type of effect can be used
in a deck. The idea for this is to allow dangerous card effects into
the game in a way that doesn't break the game.

So far, the ultra card texts I've experimentally added include:

* On-summon minion abilities.
* Enemy summon response.
* Enemy spell or ability response.
* Enemy defense response.
* HP cost.
* Self-mill cost.
* Targeting all on-field minions below a certain level.
* Targeting all attacking or defending minions below a certain level.
* Targeting cards in your grave (non-randomly).
* Searching cards in your deck.
* Targeting any card in enemy's granary.
* Direct damage.
* Countering enemy effects.

I've also added a regular "distract" effect, which removes the targets
from the attacking or defending team they were assigned to (effectively
negating their attack).

This is just a first pass. It's likely that I'll have to adjust these.