Designed the first ultra card: Lucifer. Seems a good place to start.
I slightly tweaked Lucifer to take advantage of ultra texts: it now
activates DEVIL'S BARGAIN in response to a summon (making it much
more powerful), and it also places an Undying Counter on any on-field
minion after the fact. To keep it at level 10, I removed the
down-leveling part of the ability, but in return, it now has 3 SP
instead of just 2 (matching the stats of Death).