Removed the multi-player damage effect's advantage in 3+ player games.
I realized that being able to damage all players simultaneously is
just better than damaging one enemy, because you can just pair it
with a heal. Despite that, the costs were the same.

I don't think I like the idea of instant wins in 3+ player games off
of a single card, even if it's really expensive, anyway. Better to
have to choose one player to go after at a time.