Tanked the level benefit of the self-mill cost.
The self-mill cost is intended, first and foremost, as a way to enable
decks that want to combine it with in-deck destruction activation
conditions to derive a benefit that way. It is *not* intended to
enable a level 0 card that on average will be an FTK.

At the same time, I've decreased the number of Ultra Points it costs
to use the self-mill cost from 2 to 1.

With these changes, the closest to an instant win level 0 spell would
be:

Pay 32 HP and self-destroy 48 cards on top of your deck→
8-dice-wound enemy.

This would on average inflict 28 damage, so it would be highly
inconsistent as an FTK even if it's drawn (which will also be
inconsistent because this card will cost 4 Ultra Points, meaning
only 2 copies are possible, and it requires at least a 54-card
deck). This should bring the possibility of an FTK within the
realm of enough implausibility.

Instant-win cards will still be possible of course, but they'll
require the use of resources, which will stop them from being used
immediately even if they're drawn immediately. The longer a game
goes, the less likely the player is to be able to pay a high HP
cost or a high mill cost, so this should balance itself out.