Modified the definition of hex-play to make it more worthwhile.
I've decided that hex-play will actually halve the resource cost.
My reasoning for this is it necessarily takes a 2-card combo to take
advantage of hex-play activation conditions, and that should be
rewarded somehow.

This is somewhat reminescent of the cost reductions of old editions,
but importantly, it's done during gameplay rather than at card
creation time. This somewhat limits shenanigans, and it also avoids
the problem of adding complexity to the card creation process. It's
also limited to only some of the activation conditions.