Added the "conscript" keyword.
This is a similar idea to Berzerk Counters, but instead of making
a minion attack alone, it adds it to the existing team of its
controller (forcing it to participate in a normal fight). Also,
instead of stopping its abilities from being used for the rest of
the turn, it fatigues it, and has a requirement that only minions
which are either unexhausted or not fatigued can be conscripted.
This also prevents infinite loop abuses, but in a different way.

The main reason you would use this effect is to effectively add a
bonus attacker or defender, either as a surprise or perhaps as
something like Famine (and in fact, I plan on changing Famine's
effect to conscript the pawns rather than placing Berzerk Counters).