Edited the wording of one of the common activation conditions.
There's some ambiguity with the phrasing, "When this on-field minion
is destroyed". It doesn't specify *who* gets to activate it if it
was previously controlled by a player who isn't the owner.

I've edited the wording to make it more clear that whoever controlled
the minion gets to use the ability.