Added a generic destruction condition ultra text, with a twist.
This is the only destruction condition which specifies if you *own*
it. That means it's the only destruction condition which cannot be
turned against you, and in fact, if your opponent were to steal it,
you could turn that against them. However, it's also the most flexible
self-destruction condition. Naturally, it's an ultra text.