Added Unbreakable Counters and Assassin Counters.
These are taken from back when I was brainstorming possible static
minion traits that would have a benefit with a drawback. In the end,
that idea fell thru, largely because I couldn't come up with enough
to justify the mechanic. But I just realized I can turn two of them
into counter types.