Strengthened Weakness Counters.
I don't think this buff is enough to make Weakness Counters playable.
But this buff at the very least is necessary for it to be even
remotely considered. Reducing SP by 1 is just far too little of an
effect. It's worse both than the use of Shield Counters and than the
use of Exhaustion Counters.