New card backing!
This new card backing is much more thoughtfully put together than
the old one (which was just a GIMP texture and some random font I
had installed at the time). I decided on a sort of paint splatter
esque texture, made by generating a small amount of noise at a low
resolution, smooth-scaling up, and tweaking saturation and
brightness. For the "KotC" text, I handwrote it with a brush pen in
my modern calligraphy style, made it smaller, put it in the corner,
and also added a small link to the website in the opposite corner.

In addition to just being better and putting a link to the website
on all cards, this will make cutting out cards much easier. Because
printers aren't perfectly precise, the old backing made it impossible
to use the little dotted lines provided without rendering the
backings invalid. This new design solves that problem by having a
whitespace margin around the edges, just like the front of cards.
This means you can be quite a bit more off with your cutting without
obvious marking, and keeping card sizes consistent will be the most
important consideration instead.